Check whether string is an URI with custom function
API change: the function `isURI`, testing if a string is a valid URI
with a known scheme, is exported from the Shared module, as is the set
of known `schemes`.

The new function replaces the function of the same name
from *Network.URI*, as the latter did not check whether a scheme is
well-known.  E.g. MediaWiki wikis frequently feature pages with names
like `User:John`. These links were interpreted as URIs, thus turning
internal links into global links. This is prevented by also checking
whether the scheme of a URI is frequently used (i.e. is IANA registered
or an otherwise well-known scheme).

Fixes: #2713