Name change OSX -> macOS (#3869)
* Name change OSX -> macOS

fix commit c96b64e

This commit finishes remaining osx to macOS change, as well as replacing MacOS with macOS.

The reason for the later one is because the "correct" casing of macOS is like that. Apple styles it to looks like iOS, watchOS, tvOS, etc. And unfortunately they all start with a lowercase letter, making propercasing (or even title-casing) odd.

* fix casing of Linux, UNIX, and Windows