In rendering PandocParsecError, only print input at error location...
if the source name is `source` (i.e., the top level). Otherwise
results will be misleading, since what readM does is to look
up the source position of the error in the *original* input,
which may not match the input used by parseFromString or
in parsing from an included file.

Closes #3865.  Not a great fix, maybe there's something better
that could be done, but this should at least avoid misleading
messages.