Docx writer tests: Use new golden framework
These are based off the reader tests, with some removed (where the
reader output was identical, based on different docx inputs). There
are still more to be added: especially around custom styles.

The following error has been found in the golden tests, and will be
addressed in a future commit:

- tables seem to be joined in table test output.