Docx reader: read custom styles
This will read all paragraph and character classes as divs and spans,
respectively. Dependent styles will still be resolved, but will be
wrapped with appropriate style tags. It is controlled by the `+styles`
extension (`-f docx+styles`).

This can be used in conjunction with the `custom-style` feature in the
docx writer for a pandoc-docx editing workflow. Users can convert from
an input docx, reading the custom-styles, and then use that same input
docx file as a reference-doc for producing an output docx file. Styles
will be maintained across the conversion, even if pandoc doesn't
understand them.

Without the extension:

    $ pandoc test/docx/custom-style-reference.docx -f docx -t markdown
    This is some text.

    This is text with an *emphasized* text style. And this is text with a
    **strengthened** text style.

    > Here is a styled paragraph that inherits from Block Text.

With the extension:

    $ pandoc test/docx/custom-style-reference.docx -f docx+styles -t markdown
    ::: {custom-style="FirstParagraph"}
    This is some text.
    :::

    ::: {custom-style="BodyText"}
    This is text with an
    *[[emphasized]{custom-style="Emphatic"}]{custom-style="Emphatic"}* text
    style. And this is text with a
    **[[strengthened]{custom-style="Strengthened"}]{custom-style="Strengthened"}**
    text style.
    :::

    ::: {custom-style="MyBlockStyle"}

Closes: #1843