data/templates/default.latex:  `\usepackage{iftex}` for conditionals.
     Previously, this file guessed that it was being processed by PDF(La)TeX if
it _didn't_ detect it was being processed by Lua(La)TeX or Xe(La)TeX (in order
to only invoke engine-specific macros when it was being processed by the engine
in question.)  This was done using the `ifluatex` package's `\ifluatex` macro
and the `ifxetex` package's `\ifxetex` macro.  Improve upon this logic and make
it more readable and obvious that this is what's going on inside it by:

- no longer using either of the `ifluatex` or `ifxetex` packages
- using the `iftex` package instead, as it also supplies a macro for detecting
  if a file is being processed by PDF(La)TeX
- adjusting the affected macro conditionals to account for this by:
  - replacing all appearances of an ungainly conditional macro construct with
    usages of the `\ifPDFTeX` macro newly available from that `iftex` package
  - replacing all usages of both the `ifluatex` package's `\ifluatex` macro and
    the `ifxetex` package's `\ifxetex` macro with their equivalents in the
    `iftex` package, namely `\ifLuaTeX` and `\ifXeTeX`, respectively,

in this commit.