test-pandoc-utils.lua: remove problems with missing `/bin/false`.
Previously it was assumed that the system would have `/bin/false`
and `/bin/sed`, and these tests were skipped otherwise.

On MacOS, these utilities are located in `/usr/bin`.

Fixed by just using `sed` and `false` -- these should always be
in the path.  Removed the "skipping" behavior, replaced with a
check for Windows.  On Windowns, we use `echo` and `cd`, which
should always exist.

Not yet checked on Windows.