Markdown reader: extension to insert source code position attributes
Optionally inserts source-code position attributes for `CodeBlock` in Markdown
sources. The extension is `Ext_code_block_source_position`.

This reads the parser position (`SourcePos name line col`) at the beginning of
parsing a `CodeBlock` and adds the attributes
`[("sourceName", name), ("sourceLine", line), ("sourceColumn", col)]` if the
extension is turned on.