HTML writer: Add option to put emojis in <span> tags
Adds a new `--html-emoji-class=STRING` option that will cause emojis to
be wrapped in <span> tags with the specified class.

This makes it easy to use a style sheet that has a different font for
emojis (for example a color font such as Noto Color Emoji)

Emojis are detected via a reverse lookup on the emojis map defined in
Emoji.hs. Meaning that any emoji unicode point written verbatim in the
source document and that isn't defined in Emoji.hs will not be wrapped.

Closes: #4743