Wrap emojis in Span
Emojis (as supported in Markdown and CommonMark readers, e.g ":smile:")
were simply translated into the corresponding unicode code point.

This commit changes so that they are also wrapped in a `Span` node. This
makes it easy to do special handling such as giving them a special font
in HTML output.

The wrapping span has class set to `emoji` and a `data-emoji` attribute
with the name of the emoji.

Closes #4743.