Fix regression finding templates in user data directory.
Under version 2.2.1 and prior pandoc found latex templates in the
templates directory under the data directory, but this no longer
works in 2.2.2.

MANUAL says: "If the template is not found, pandoc will search for it in
the templates subdirectory of the user data directory (see `--data-dir`)."

This commit fixes the regression, which stems from 07bce91.

Closes #4777.