MANUAL: add anchors to option descriptions
This is useful for linking to option descriptions in the manual, e.g. on
the mailing list, Stack Exchange etc.

The anchors are implemented as a span wrapping each option signature
line with and id attribute of the form `#option-<option-name>` where
`<option-name>` is the *first* long option name on the line minus the
leading double dash, e.g.

    [`-t` *FORMAT*, `-w` *FORMAT*, `--to=`*FORMAT*, `--write=`*FORMAT*]{#option-to}

I must confess that I didn't put a lot of work into the edit itself; I
just ran the following Perl substitution command on the *Options*
section:

    s/^((?=`--?\w[-\w]*).*?`--(\w[-\w]+).*)/[$1]{#option-$2}/

Due to the layout of the *Options* section this should do it correctly.

I have looked through the diff and it seems like everything is as
expected.

I have also checked that there are no duplicate matches against the
regex `#option-[-\w]+`.

I haven't added any links to these anchors from elsewhere in the manual
at this time, but plan to do so if and when this PR is accepted.